fix: chapter titles wrongly being in English
Happens when:
1. There are no preexisting chapters
2. The model is gpt-3.5-turbo-16k
3. The requested language is non-English

It's wrong when the language is different from English:

> System: output summary in language "es"
> user: <transcript>
> assistant: <summary in "en">

We give the wrong example of the assistant's response where it disobeys
the language direction, so it might disobey in when it gets output for
the actual transcript.

Fixes https://github.com/mthli/better-youtube-summary-server/issues/2

Though it only affects chapter titles and not their contents due to
another bug where we regenerate chapter contents in `summarize()`
after `_generate_multi_chapters` with 16K model.